personal/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-modeltranslation==0.18.11
django-preventconcurrentlogins==0.8.2
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
